# src/main/kotlin/eventichs/api/eventichs_api/Controleurs/openapi.yaml
openapi: 3.0.3
info:
  description: Eventichs_API API
  title: Eventichs_API API
  version: 1.0.0
servers:
- url: https://Eventichs_API
paths:
...
  /<unknown>/invitations/{id}:
    get:
      description: Obtenir la liste des invitations par l'id de l'invité. dans le cas ou l'id représente une organisation, Obtenir la liste des demandes d'invitations.
      operationId: obtenirInvitationsParIdUtilisateur
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: Retourne la liste des invitations.
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: Utilisateur non trouvé.
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: Syntaxe incorrecte.
      summary: GET ${..}/invitations/{id}
      x-accepts: '*/*'
  /<unknown>/invitations:
    post:
      description: Ajouter une invitation ou une demande d'invitation, dépendamment du type de l'invitateur et de l'invité.
      operationId: inviterOuDemanderInvitation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invitation'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: L'invitation est créée, et on obtient l'objet invitation créé.
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: l'invité ou l'inviteur n'existe pas.
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: l'invité fait déja partie de l'évènement ou de l'organisation, ou l'évènement provient d'un groupe dont l'invité n'a pas accès.
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: syntaxe incorrecte.
      summary: POST ${..}/invitations
      x-contentType: application/json
      x-accepts: '*/*'
  /<unknown>/invitation/{id}:
    delete:
      description: Supprimer une invitation.
      operationId: supprimerInvitation
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: Invitation supprimée
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: Invitation non-trouvée
      summary: DELETE ${..}/invitation/{id}
      x-accepts: '*/*'
    put:
      description: Répondre à une invitation ou demande.
      operationId: majInvitationGroupe
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: Réponse enregistrée
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: syntaxe invalide
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: Invitation non trouvée
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntity'
          description: Invitation déja répondue
      summary: PUT ${..}/invitation/{id}
      x-contentType: application/json
      x-accepts: '*/*'
components:
  schemas:
...
    ResponseEntity:
      example:
        headers: '{}'
      properties:
        headers:
          properties: {}
          type: object
      type: object
...
    Invitation:
      example:
        codeExpediteur: 6
        codeDestinataire: 1
        id: 0
        codeQuoiRejoindre: 5
        status: status
      properties:
        id:
          format: int32
          type: integer
        codeExpediteur:
          format: int32
          type: integer
        codeDestinataire:
          format: int32
          type: integer
        codeQuoiRejoindre:
          format: int32
          type: integer
        status:
          type: string
      type: object